lib_airnav::parsers::arinc424::definitions

Module enums

ARINC 424 Definitions - Enums

This module contains the enums for the ARINC 424 data. Enums are not a formally defined data type in the spec, but are obviously represented as such given the character mapping tables.

Example is 5.4 - Section Code which describes the major section of the record.

Enums

AirportHeliportApproachRouteType

AirportSubsection

AirspaceSubsection

BoundaryCode

CompanyRoutesSubsection

CrossingAltitudeDescription

CustomerAreaCode

EnrouteAirwayRouteType

EnrouteSubsection

HeliportSubsection

Level

MORASubsection

NavaidSubsection

PreferredRouteRouteType

RecordType

SIDRouteType

STARRouteType

Section

TablesSubsection

TurnDirection

TurnDirectionValid