lib_airnav::parsers::arinc424::definitions

Module identifiers

ARINC 424 Definitions - Identifiers

This module contains the identifiers for the ARINC 424 data. Identifiers are simple alphanumeric string fields. The fields may or may not be constrained to be exactly the specified length always. Conversely, some fields are minimized as needed in the spec to accomodate record layouts.

Example is 5.6 - Airport/Heliport Identifier which describes a 4-character code that identifies an airport or heliport.

Structs

LengthLimitedIdentifier

Type Aliases

AirportHeliportIdentifier

5.6 Airport/Heliport Identifier

ApproachRouteIdentifier

5.10 Approach Route Identifier

ContinuationRecordNumber

DMEIdentifier

5.38 DME Identifier

EnrouteRouteIdentifier

5.8(A) Enroute Route Identifier

FixIdentifier

5.13 Fix Identifier

IcaoCode

5.14 ICAO Code

InboundCourseTheta

PathAndTermination

PreferredRouteIdentifier

5.8(B) Preferred Route Identifier

RecommendedNavaid

RegionCode

5.40 Region Code

SidStarRouteIdentifier

5.9

TransitionIdentifier

5.11 Transition Identifier

VORNDBIdentifier

5.33 VOR/NDB Identifier

WaypointDescriptionCode