lib_airnav::parsers::arinc424::definitions

Module numerics

ARINC 424 Definitions - Numerics

This module contains the numeric types for the ARINC 424 data. Numerics are simple numeric fields. There are several cases we need to handle in the spec, but are more nice to work with when converted appropriately. Latitudes and Longitudes are such cases where we convert to a more readable format.

Combined Numerics

Some fields encode multiple diverging types of numeric values. For example, 5.27 - Route Distance From encodes both a time and a distance.

To handle this, there are specific fields that capture the union of the two types appropriately.

Structs

Altitude

FloatNumeric

IntNumeric

LatitudeNumeric

LongitudeNumeric

MagneticVariationNumeric

UintNumeric

Enums

AltitudeMinimumAltitude

TimeDistanceNumeric

Type Aliases

CycleDate

FileRecordNumber

InboundCourse

Latitude

5.36 Latitude

Longitude

5.37 Longitude

MagneticVariation

5.39 Magnetic Variation

OutboundCourse

Rho

RouteDistanceFrom

SequenceNumber

Theta

VORNDBFrequency

5.34 VOR/NDB Frequency