lib_airnav::parsers::arinc424

Module parser

ARINC 424 Parser

This module contains the parser for the ARINC 424 data. It is a wrapper around the various revisions we support in the package for processing data. This allows for easy revision switching without making us dynamically allocate a bunch more heap objects to handle dynamicity within a trait-based approach, so while it might be a bit more verbose, it is more efficient.

Enums

Arinc424Parser

Arinc424Version

Arinc424VersionedRecord