lib_airnav::parsers::arinc424::records

Module enroute_airway_restriction

Structs

EnrouteAirwayRestrictionAltitudeExclusionFormattedTimeContinuationRecord

4.1.21.A.3 Enroute Airway Restriction Altitude Exclusion Formatted Time Continuation Record

EnrouteAirwayRestrictionAltitudeExclusionNarrativeTimeContinuationRecord

4.1.21.A.4 Enroute Airway Restriction Altitude Exclusion Narrative Time Continuation Record

EnrouteAirwayRestrictionAltitudeExclusionPrimaryExtensionContinuationRecord

4.1.21.A.2 Enroute Airway Restriction Altitude Exclusion Primary Extension Continuation Record

EnrouteAirwayRestrictionAltitudeExclusionPrimaryRecord

4.1.21.A.1 Enroute Airway Restriction Altitude Exclusion Primary Record

EnrouteAirwayRestrictionCruisingTableReplacementFormattedTimeContinuationRecord

4.1.21.D.2 Enroute Airway Restriction Cruising Table Replacement Formatted Time Continuation Record

EnrouteAirwayRestrictionCruisingTableReplacementNarrativeTimeContinuationRecord

4.1.21.D.3 Enroute Airway Restriction Cruising Table Replacement Narrative Time Continuation Record

EnrouteAirwayRestrictionCruisingTableReplacementPrimaryRecord

4.1.21.D.1 Enroute Airway Restriction Cruising Table Replacement Primary Record

EnrouteAirwayRestrictionNoteRestrictionContinuationRecord

4.1.21.B.2 Enroute Airway Restriction Note Restriction Continuation Record

EnrouteAirwayRestrictionNoteRestrictionPrimaryRecord

4.1.21.B.1 Enroute Airway Restriction Note Restriction Primary Record

EnrouteAirwayRestrictionSeasonalClosureFormattedTimeContinuationRecord

4.1.21.C.3 Enroute Airway Restriction Seasonal Closure Formatted Time Continuation Record

EnrouteAirwayRestrictionSeasonalClosureNarrativeTimeContinuationRecord

4.1.21.C.4 Enroute Airway Restriction Seasonal Closure Narrative Time Continuation Record

EnrouteAirwayRestrictionSeasonalClosurePrimaryRecord

4.1.21.C.1 Enroute Airway Restriction Seasonal Closure Primary Record