lib_airnav::parsers::arinc424::rev18::definitions

Module identifiers

ARINC 424 Definitions - Identifiers

This module contains the identifiers for the ARINC 424 data. Identifiers are simple alphanumeric string fields. The fields may or may not be constrained to be exactly the specified length always. Conversely, some fields are minimized as needed in the spec to accomodate record layouts.

Example is 5.6 - Airport/Heliport Identifier which describes a 4-character code that identifies an airport or heliport.

Enums

GridMora

5.143 Grid MORA

Type Aliases

AirportHeliportIdentifier

5.6 Airport/Heliport Identifier

AirwayRestrictionIdentifier

5.154 Airway Restriction Identifier

AirwayRestrictionNotes

5.163 Airway Restriction Notes

AirwayRestrictionStartEndDate

5.157 Airway Restriction Start/End Date

ApproachRouteIdentifier

5.10 Approach Route Identifier

ApproachTypeIdentifier

5.262 Approach Type Identifier

AtaIataDesignator

5.107 ATA/IATA Designator

AtcIdentifier

5.236 ATC Identifier

Callsign

5.105 Callsign Note: This is specced as being 25 characters long, but the record for comms shows 30 characters. Thus we change this to 30 characters to match the record.

CenterFix

5.144 Center Fix

CompanyRouteIdent

5.76 Company Route Ident

CompanyRouteToFix

5.83(A) To Fix for Company Routes

ContinuationRecordNumber

5.16 Continuation Record Number

ControlledAirspaceCenter

5.214 Controlled Airspace Center

ControlledAirspaceClassification

5.215 Controlled Airspace Classification

ControlledAirspaceName

5.216 Controlled Airspace Name

ControllingAgency

5.140 Controlling Agency

DMEIdentifier

5.38 DME Identifier

DatumCode

5.197 Datum Code

EnrouteAlternateAirportHeliport

5.148 Enroute Alternate Airport/Heliport

EnrouteRouteIdentifier

5.8(A) Enroute Route Identifier

EnrouteTransition

5.85 Enrt Transition

FinalApproachSegmentDataCrcRemainder

5.229 Final Approach Segment Data CRC Remainder

FirRdoIdentifier

5.190 FIR/RDO Identifier

FirUirAddress

5.151 FIR/UIR Address

FirUirIdentifier

5.116 FIR/UIR Identifier

FirUirName

5.125 FIR/UIR Name

FixIdentifier

5.13 Fix Identifier

FromToAirportHeliportFix

5.75 From/To Airport/Heliport/Fix (FROM/TO AIRPORT/HELIPORT/FIX)

GateIdentifier

5.56 Gate Identifier (GATE IDENT)

GenericSubsection

5.5 Generic subsection code

GeographicalEntity

5.219 Geographical Entity

GeographicalReferenceTableIdentifier

5.218 Geographical Reference Table Identifier

GlsStationIdentifier

5.243 GLS Station Identifier

GlsTdmaSlots

5.246 GLS TDMA Slots

IcaoCode

5.14 ICAO Code

InitialTerminusFixOrAirport

5.194 Initial/Terminus Fix or Airport

LevelOfServiceName

5.275 Level of Service Name

LocalizerMlsGlsIdentifier

5.44 Localizer/MLS/GLS Identifier (LOC, MLS, GLS IDENT)

MarkerCode

5.111 Marker Code (Morse)

MarkerIdentifier

5.110 Marker Identifier (Enroute Marker) (IDENT)

MoraStartingLatitude

5.141 MORA Starting Latitude

MoraStartingLongitude

5.142 MORA Starting Longitude

MultipleCode

5.130 Multiple Code

Name

5.60 Name (NAME), Gate and Holding Pattern records

NameOfFacility

5.71 Name Field, Navaid/Airport/Heliport/Enroute Marker records

Notes

5.61 Notes (Continuation Records) (NOTES)

PadIdentifier

5.180 Pad Identifier

PathAndTermination

5.21 Path and Termination

PreferredRouteIdentifier

5.8(B) Preferred Route Identifier

PreferredRouteToFix

5.83(B) To Fix for Preferred Routes

PrimaryAndAdditionalAlternateIdentifier

5.253 Primary and Additional Alternate Identifier

ProcedureDescription

5.237 Procedure Description

RecommendedNavaid

5.23 Recommended NAVAID

ReferencePathIdentifier

5.257 Reference Path Identifier

RegionCode

5.41 Region Code

RemoteFacility

5.200 Remote Facility

RemoteSiteName

5.189 Remote Site Name

RestrictiveAirspaceDesignation

5.129 Restrictive Airspace Designation

RestrictiveAirspaceName

5.126 Restrictive Airspace Name

RouteIndicator

5.224 Route Indicator

RunwayDescription

5.59 Runway Description (RUNWAY DESCRIPTION)

RunwayIdentifier

5.46 Runway Identifier (RUNWAY ID)

RunwayTransition

5.84 Runway Transition

SectorFacility

5.185 Sector Facility

SectorizationNarrative

5.186 Sectorization Narrative

SidStarApproachAirwayIdentifier

5.78 SID/STAR/App/AWY (S/S/A/AWY), SID/STAR/AWY (S/S/AWY)

SidStarRouteIdentifier

5.9

StartEndDate

5.153 Start/End Date

TaaWaypoint

5.273 TAA Waypoint

TerminalAlternateAirport

5.87 Terminal/Alternate Airport

TimeOfOperation

5.195 Time of Operation

TransitionIdentifier

5.11 Transition Identifier

VORNDBIdentifier

5.33 VOR/NDB Identifier

WaypointDescriptionCode

5.17 Waypoint Description Code Note: This should be a complex type, but the rules are a bit difficult to parse out atm TODO: come back to this

WaypointNameDescription

5.43 Waypoint Name/Description (NAME/DESC)