lib_airnav::parsers::arinc424::rev18_faa::definitions

Module complexes

ARINC 424 Definitions - Complexes

This module contains the complex types for the ARINC 424 data. This is not a formally defined thing, but rather a term I have used to describe the encoding of multiple annotations onto one field definition in the spec The first example of which is given in 5.35 - Navaid Class which describes a 5-letter code that adds annotations onto a Navaid record.

The way we encode these complexes are as tuples or nested enums. In the above example, the Navaid Class is encoded as a tuple of 5 enum fields. Thus, when parsing a VHF Navaid record in particular we represent this as the combination of:

VHFNavaidType1

VHFNavaidType2

VHFRangePower

VHFAdditionalInfo

Collocation

Structs

AircraftUseGroupIndicator

5.221 Aircraft Use Group Indicator

AirportHeliportCommunicationsServiceIndicator

5.106(A) Airport Heliport Communications Service Indicator

BoundaryVia

5.118 Boundary Via

CircularPadDimensions

DuplicateIndicator

5.114 Duplicate Indicator

EnrouteCommunicationsServiceIndicator

5.106(B) Enroute Communications Service Indicator

FacilityCharacteristics

5.93 Facility Characteristics (FAC CHAR)

GLSStationType

5.247 GLS Station Type

LegTypeCode

5.238 Leg Type Code

MarkerLocatorNavaidClass

5.35(C) Localizer Marker/Locator Navaid Class

MarkerType

5.99 Marker Type

NDBNavaidClass

5.35(B) NDB Navaid Class

NameFormat

5.196 Name Format Indicator

NavaidAltitudeLimitation

5.209 Navaid Altitude Limitation

NavaidDistanceLimitation

5.208 Navaid Distance Limitation

NumberOfEnginesRestriction

5.232 Number of Engines Restriction

PreferredRouteUseIndicator

5.220 Preferred Route Use Indicator

RectangularPadDimensions

SectorBearing

5.146 MSA Sector Bearing

SectorFromTo

5.207 Sector From / Sector To Note: This should resolve to degrees, but for now we will keep it as is

TaaSectorRadius

5.274 TAA Sector Radius

Timezone

5.178 Timezone

VHFNavaidClass

5.35(A) NAVAID Class

WaypointType

5.42 Waypoint Type (TYPE)

WaypointUsage

5.82 Waypoint Usage

Enums

AircraftUseGroup

AirportHeliportCommunicationsServiceIndicator1

AirportHeliportCommunicationsServiceIndicator2

AirportHeliportCommunicationsServiceIndicator3

BoundaryViaEndPoint

BoundaryViaPathType

Collocation

CruiseLevelFromTo

5.136 Cruise Level From/To Note: see 5.121

DescriptiveAltitudeLimits

DuplicateIndicatorAirspace

EnrouteCommunicationsServiceIndicator1

EnrouteCommunicationsServiceIndicator2

EnrouteCommunicationsServiceIndicator3

FacilityCharacteristicsEmissionType

FacilityCharacteristicsFreqAndBackcourse

FacilityCharacteristicsRepetitionAndCollocation

FacilityCharacteristicsSynchronicity

FacilityCharacteristicsVoiceIdent

GLSStationType1

GLSStationType2

GLSStationType3

LegTypePath

LegTypeTurnIndication

LowerUpperLimit

5.121 Lower/Upper Limit Note: maybe belongs in the common numeric field section? unsure

MarkerCollocation

MarkerTypeLocation

MarkerTypeLocator

MarkerTypeMarker

MaximumAltitude

5.127 Maximum Altitude Note: see 5.121

NDBAdditionalInfo

NDBCollocation

NDBNavaidType1

NDBNavaidType2

NDBRangePower

NameFormatType1

NameFormatType2

NumberOfEnginesRestrictionIndicator

PadDimensions

5.176 Pad Dimensions

PreferredRouteRNAVRequirement

PreferredRouteType

PrimaryWaypointType

PublishedUse

SecondaryWaypointType

TimezoneZone

VHFAdditionalInfo

VHFNavaidType1

VHFNavaidType2

VHFRangePower

WaypointUsageAltitude

WaypointUsageRNAVIndicator

Type Aliases

CommunicationsSectorization

5.183 Comms Sectorization

DuplicateIndicatorDuplicates