lib_airnav::parsers::arinc424::rev18_faa::definitions

Module validators

Functions

is_valid_altitude_description

Validates the altitude description to ensure we don’t have values that are invalid based on the record type

is_valid_boundary_code

Validates the area code and boundary code combination to ensure they are a possible combination that is allowed. Note that this should only check boundaries, that is, bordering regions that can touch

is_valid_reference

Validates the reference, section, and subsection combination to ensure they are a possible combination that is allowed.

is_valid_route_type_and_qualifier_combination

Validates the route type and qualifier combinations to ensure they are a possible combination that is allowed for an airport heliport approach route.