lib_airnav::parsers::arinc424::rev23::definitions

Module numerics

ARINC 424 Definitions - Numerics

This module contains the numeric types for the ARINC 424 data. Numerics are simple numeric fields. There are several cases we need to handle in the spec, but are more nice to work with when converted appropriately. Latitudes and Longitudes are such cases where we convert to a more readable format.

Combined Numerics

Some fields encode multiple diverging types of numeric values. For example, 5.27 - Route Distance From encodes both a time and a distance.

To handle this, there are specific fields that capture the union of the two types appropriately.

Structs

FinalApproachCourseAsRunway

5.299 Final Approach Course as Runway

Enums

CommunicationsFrequency

5.103 Communications Frequency

PathPointTCH

5.265 Path Point TCH

SectorAltitude

5.147 Sector Altitude

Type Aliases

ASDA

5.315 ASDA

AirportHeliportElevation

5.55 Airport/Heliport Elevation (ELEV)

AirwayRestrictionAltitude

5.161 Airway Restriction Altitude

AlongTrackDistance

5.231 Along Track Distance

AlternateDistance

5.88 Alternate Distance

ArcBearing

5.120 Arc Bearing

ArcDistance

5.119 Arc Distance

ArcRadius

5.204 ARC Radius

CirclingCategoryDistance

5.292 Circling Category Distance

CommunicationsAltitude

5.184 Communication Altitude

CommunicationsDistance

5.188 Communications Distance

ComponentElevation

5.74 Component Elevation (GS ELEV, EL ELEV, AZ ELEV, BAZ ELEV, GLS ELEV)

CostIndex

5.89 Cost Index

CourseWidthAtThreshold

5.228 Course width at threshold

CruiseAltitude

5.86 Cruise Altitude (CRUISE ALTITUDE)

CruiseTableCourseFromTo

5.135 Course From/To (Cruise Table)

CycleDate

5.32 Cycle Date

DMEElevation

5.40 DME Elevation

DistanceToAlternate

5.251 Distance To Alternate

FacilityElevation

5.92 Facility Elevation (FAC ELEV)

FileRecordNumber

5.31 File Record Number

FixedRadiusTransitionIndicator

5.254 Fixed Radius Transition Indicator

FlightPlanningAltitude

5.240 Flight Planning Altitude

FrequencyProtectionDistance

5.150 Frequency Protection Distance

GLSServiceVolumeRadius

5.245 GLS Service Volume Radius

GLSWgs84StationElevation

5.248 GLS WGS84 Station Elevation

GlideslopeAngle

5.52 Glideslope angle (GS ANGLE) Minimum Elevation Angle (MIN ELEV ANGLE)

GlideslopeBeamWidth

5.96 Glideslope Beam Width

GlideslopePosition

5.50 Glideslope Position (GS FR RW THRES) Elevation Position (EL FR RW THRES)

HelicopterProcedureCourse

5.269 Helicopter Procedure Course

HelipadMaximumRotorDiameter

5.321 Helipad Maximum Rotor Diameter

HeliportIdentifierOrientation

5.324 Heliport Identifier Orientation

HeliportOrientation

5.323 Heliport Orientation

HighFrequencyCommunicationsFrequency

HighPrecisionLatitude

5.267 High Precision Latitude

HighPrecisionLongitude

5.268 High Precision Longitude

HoldingPatternMagneticVariation

5.343 Holding Pattern Magnetic Variation

HoldingSpeed

5.175 Holding Speed

HorizontalAlertLimit

5.263 Horizontal Alert Limit

IlsDmeBias

5.90 ILS/DME Bias

InboundCourse

5.28 Inbound Course

InboundHoldingCourse

5.62 Inbound Holding Course (IB HOLD CRS)

LDA

5.316 LDA

LandingThresholdElevation

5.68 Landing Threshold Elevation (LANDING THRES ELEV)

Latitude

5.36 Latitude

LegLength

5.64 Leg Length (LEG LENGTH)

LegTime

5.65 Leg Time (LEG TIME)

LocalizerBearing

5.47 Localizer Bearing (LOC BRG)

LocalizerFrequency

5.45 Localizer Frequency (FREQ)

LocalizerPosition

5.48 Localizer Position (LOC FR RW END / AZ/BAZ FR RW END) Azimuth/Back Azimuth Position (AZ/BAZ FR RW END)

LocalizerWidth

5.51 Localizer Width (LOC WIDTH)

LongestRunway

5.54 Longest Runway (LONGEST RWY)

Longitude

5.37 Longitude

MLSAzimuthBearing

5.167 MLS Azimuth/Back Azimuth Bearing

MLSAzimuthCoverageSector

5.172 MLS Azimuth/Back Azimuth Coverage Sector

MLSAzimuthProportionalAngle

5.168 MLS Azimuth/Back Azimuth Proportional Angle

MLSChannel

5.166 MLS Channel

MLSElevationAngleSpan

5.169 MLS Elevation Angle Span

MLSNominalElevationAngle

5.173 MLS Nominal Elevation Angle

MagneticVariation

5.39 Magnetic Variation

MaximumAllowableHelicopterWeight

5.309 Maximum Allowable Helicopter Weight

MinimumAltitude

5.30 Altitude / Minimum Altitude

MinorAxisBearing

5.100 Minor Axis Bearing

MsaRadiusLimit

5.145 MSA Radius Limit

NDBFrequency

5.34(B) NDB Frequency

OrthometricHeight

5.227 Orthometric height

OutboundCourse

5.26 Outbound Course

PathPointTCHFeet

PathPointTCHMeters

PreferredApproachBearing

5.325 Preferred Approach Bearing

ProcedureDesignMagneticVariation

5.290 Procedure Design Magnetic Variation

RNPLevelOfService

5.296 RNP Level of Service

RVSMMaximumLevel

5.295 RVSM Maximum Level

RVSMMinimumLevel

5.294 RVSM Minimum Level

ReferencePathDataSelector

5.256 Reference Path Data Selector

RequiredNavigationPerformance

5.211 Required Navigation Performance

Rho

5.25 Rho

RouteDistanceFrom

5.27 Route Distance From

RunwayBearing

5.58 Runway Bearing (RWY BRG)

RunwayGradient

5.212 Runway Gradient

RunwayLength

5.57 Runway Length (RUNWAY LENGTH)

RunwayStarterExtension

5.312 Runway Starter Extension

RunwayWidth

5.109 Runway Width

SBASFinalApproachCourse

5.320 SBAS Final Approach Course

SBASGBASChannel

5.244 SBAS/GBAS Channel

SBASGBASGlidePathAngle

5.226 SBAS/GBAS Glide Path Angle

SBASGBASLengthOffset

5.259 SBAS/GBAS Length Offset

SequenceNumber

5.12 Sequence Number

SpecialActivityAreaSize

5.280 Special Activity Area Size

SpeedLimit

5.72 Speed Limit (SPEED LIMIT)

SpeedLimitAltitude

5.73 Speed Limit Altitude (SPEED LIMIT ALTITUDE)

StationDeclination

5.66 Station Declination (STN DEC)

Stopway

5.79 Stopway (STOPWAY)

TODA

5.314 TODA

TORA

5.313 TORA

TerminalProcedureFlightPlanningLegDistance

5.260 Terminal Procedure Flight Planning Leg Distance

Theta

5.24 Theta

ThresholdCrossingHeight

5.67 Threshold Crossing Height (TCH)

ThresholdDisplacementDistance

5.69 Threshold Displacement Distance (DSPLCD THR)

TouchdownZoneElevation

5.97 Touchdown Zone Elevation

TransitionAltitudeLevel

5.53 Transition Altitude/Level (TRANS ALTITUDE/LEVEL)

TrueBearing

5.94 True Bearing

UltraHighFrequencyCommunicationsFrequency

VORFrequency

5.34(A) VOR Frequency

VerticalAlertLimit

5.264 Vertical Alert Limit

VerticalAngle

5.70 Vertical Angle (VERT ANGLE)

VerticalScaleFactor

5.293 Vertical Scale Factor

VerticalSeparation

5.137 Vertical Separation

VeryHighFrequencyCommunicationsFrequency

WGS84EllipsoidHeight

5.225 WGS-84 ellipsoid height