lib_airnav::parsers::arinc424::types

Module fields

ARINC 424 Field Parsers

This module contains the core type converters for the ARINC 424 field types.

Not all types can be used commonly, so we allow version specifications to derive their own in that revision package.

Later improvements could be to find more commonality but it won’t be prio’d now since it generally will get things forward in a stable manner without worry about breakage across revisions.

Structs

AltitudeNumeric

FieldParseError

FloatNumeric

IntNumeric

LatitudeNumeric

LengthLimitedIdentifier

LongitudeNumeric

MagneticVariationNumeric

UintNumeric

VariableFloatNumeric

Enums

BearingNumeric

Bearing Numeric at first glance seems similar to Declination Numeric, but in this case the indicator for True is the only one to look for and is at the end of the byte array passed in.

DeclinationNumeric

Declination Numeric at first glance seems similar to Magnetic Variation, but is actually used in much more different contexts that need different semantic handling of values.

MinimumAltitudeNumeric

MultiUnitAltitudeNumeric

TimeDistanceNumeric

Constants

BLANK

Traits

ParseableField

Functions

coalesce_into_number