lib_airnav::parsers::arinc424::definitions

Module enums

ARINC 424 Definitions - Enums

This module contains the enums for the ARINC 424 data. Enums are not a formally defined data type in the spec, but are obviously represented as such given the character mapping tables.

Example is 5.4 - Section Code which describes the major section of the record.

Enums

ATCWeightCategory

5.235 ATC Weight Category

ATNATSUGroundFacilityUseIndicator

5.337 ATN ATSU Ground Facility Use Indicator

AirportHeliportApproachRouteType

5.7(E) Airport Heliport Approach Route Type

AirportSubsection

5.6(C) Airport Subsection Code

AirspaceLimitUnitIndicator

5.133 Unit Indicator

AirspaceSubsection

5.6(F) Airspace Subsection Code

AirwayRestrictionAltitudeUnit

5.160 Units of Altitude for airway restriction

AlternateRecordType

5.250 Alternate Record Type

AlternateType

5.252 Alternate Type

AltitudeExclusionIndicator

5.202 Exclusion Indicator

AtcAssignedOnly

5.159 ATC Assigned Only

AtcIndicator

5.81 ATC Indicator (ATC)

BaroVnavAuthorization

5.155 BARO-VNAV Authorization

BlockAltitudeIndicator

5.203 Block Indicator

BoundaryCode

5.18 Boundary Code

CommunicationsClass

5.283 Communications Class

CommunicationsType

5.101 Communications Type

CommunicationsTypeRecognizedBy

5.287 Communications Type Recognized By

CommunicationsUsedOn

5.289 Communications Used On

CompanyRouteVIACode

5.77(A) Company Route VIA Code

CompanyRoutesSubsection

5.6(D) Company Routes Subsection Code

ComponentAffectedIndicator

5.206 Component Affected Indicator

ContinuationRecordApplicationType

5.91 Continuation Record Application Type (APPL)

ContinuationRecordTimeCode

5.131(B) Continuation Record Time Code

ControlledAirspaceIndicator

5.217 Controlled Airspace Indicator

ControlledAirspaceType

5.213 Controlled Airspace Type

CrossingAltitudeDescription

5.29 Crossing Altitude Description

CruiseTableIdentifier

5.134 Cruise Table Identifier

CustomerAreaCode

5.8 Customer/Area Code

DMEExpandedServiceVolume

5.339 DME Expanded Service Volume

DMEOperationalServiceVolume

5.277 DME Operational Service Volume

DaylightTimeObservedIndicator

5.179 Daylight Time Observed Indicator

DistanceDescription

5.187 Distance Description

ElevationType

5.98 Elevation Type

EnrouteAirwayRestrictionFlag

5.164 Enroute Airway Restriction Flag

EnrouteAirwayRouteType

5.7(A) Enroute Airway Route Type

EnrouteDirectionalRestriction

5.115(A) Enroute Directional Restriction

EnrouteMarkerPower

5.113 High/Low (Enroute Marker)

EnrouteSubsection

5.6(A) Enroute Subsection Code

FIRFRATransitionType

5.311 FIR/FRA Transition Type

FirUirEntryReport

5.124 FIR/UIR Entry Report

FirUirIndicator

5.117 FIR/UIR Indicator

FirUirReportingUnitsAltitude

5.123 FIR/UIR Reporting Units Altitude

FirUirReportingUnitsSpeed

5.122 FIR/UIR Reporting Units Speed

FixRelatedTransitionCode

5.241 Fix Related Transition Code

FrequencyUnits

5.104 Frequency Units

GBASApproachPerformanceDesignator

5.258 GBAS Approach Performance Designator

GBASOperationType

5.223(B) GBAS Operation Type

GnssFmsIndicator

5.222 GNSS/FMS Indicator

GovernmentSource

5.95 Government Source

H24Indicator

5.181 H24 Indicator

HelicopterPerformanceRequirement

5.310 Helicopter Performance Requirement

HelipadShape

5.303 Helipad Shape

HelipadType

5.322 Helipad Type

HeliportSubsection

5.6(B) Heliport Subsection Code

HeliportType

5.305 Heliport Type

HoldingPatternCourseReversalLegIndicator

5.298 Holding Pattern/Race Track Course Reversal Leg Inbound/Outbound Indicator

IfrCapability

5.108 IFR Capability

IlsMlsGlsCategory

5.80 ILS/MLS/GLS Category (CAT)

LandingThresholdElevationAccuracyComplianceFlag

5.319 Landing Threshold Elevation Accuracy Compliance Flag

Level

5.19 Level

LevelOfServiceAuthorized

5.276 Level of Service Authorized

LocalizerAzimuthPositionReference

5.49 Localizer Azimuth Position Reference (@, +, -)

MORASubsection

5.5(A) MORA Subsection Code

MagneticTrueIndicator

5.165 Magnetic/True Indicator

MarkerRadiationShape

5.112 Marker Radiation Shape

Modulation

5.197 Modulation

MultiSectorIndicator

5.286 Multi-Sector Indicator

NavaidLimitationCode

5.205 Navaid Limitation

NavaidSubsection

5.5(B) Navaid Subsection Code

NavaidUsableRange

5.149 NavaidUsableRange

NotamFlag

5.132 NOTAM Flag

PreferredMultipleApproachIndicator

5.306 Preferred Multiple Approach Indicator

PreferredRouteDirectionalRestriction

5.115(B) Preferred Route Directional Restriction

PreferredRouteRouteType

5.7(B) Preferred Route Route Type

PreferredRouteVIACode

5.77(B) Preferred Route VIA Code

PrimaryRecordTimeCode

5.131(A) Primary Record Time Code

ProcedureCategory

5.242 Procedure Category

ProcedureDesignAircraftCategoryOrType

5.299 Procedure Design Aircraft Category or Type

ProcedureDesignMagneticVariationIndicator

5.291 Procedure Design Magnetic Variation Indicator

ProcedureTurn

5.271 Procedure Turn

ProcedureType

5.230 Procedure Type

PublicMilitaryIndicator

5.177 Public/Military Indicator

RNAVFlag

5.234 RNAV Flag

Radar

5.102 Radar

RecordType

RemoteAltimeterFlag

5.308 Remote Altimeter Flag

ReportingCode

5.239 Reporting Code

RestrictionRecordType

5.201 Restriction Record Type

RestrictiveAirspaceLinkContinuation

5.174 Restrictive Airspace Link Continuation

RestrictiveAirspaceType

5.128 Restrictive Airspace Type

RouteInappropriateNavaidIndicator

5.297 Route Inappropriate Navaid Indicator

RunwayAccuracyComplianceFlag

5.318 Runway Accuracy Compliance Flag

RunwaySurfaceCode

5.249 Runway Surface Code

RunwayUsageIndicator

5.317 Runway Usage Indicator

SBASOperationType

5.223(A) SBAS Operation Type

SIDRouteType

5.7(C) SID Route Type

STARRouteType

5.7(D) STAR Route Type

SbasServiceProviderIdentifier

5.255 SBAS Service Provider Identifier

Section

5.4 Section Code

SequenceEndIndicator

5.210 Sequence End Indicator

SignalEmission

5.199 Signal Emission

SourceOfLALVAL

5.342 Source of LAL/VAL

SpecialActivityType

5.278 Special Activity Type

SpecialProcedureIndicator

5.307 Terminal Procedure Special Indicator

SpeedLimitDescription

5.261 Speed Limit Description

StepClimbIndicator

5.162 Step Climb Indicator

SurfaceType

5.299 Surface Type

TCHUnitsIndicator

5.266 TCH Units Indicator

TaaSectorIdentifier

5.272 TAA Sector Identifier

TablesSubsection

5.6(E) Tables Subsection Code

TchValueIndicator

5.270 TCH Value Indicator

TerminalProcedureForMilitaryIndicator

5.341 Terminal Procedure For Military Indicator

TimeIndicator

5.138 Time Indicator

TurbopropJetIndicator

5.233 Turboprop/Jet Indicator

Turn

5.63 Turn (TURN) for Holding Pattern records (TODO: investigate deduplication with 5.20 Turn Direction)

TurnDirection

5.20 Turn Direction

TurnDirectionValid

5.22 Turn Direction Valid

UnmannedAerialVehicleOnly

5.340 Unmanned Aerial Vehicle (UAV) Only

VFRCheckpointFlag

5.158 VFR Checkpoint Flag

VhfNavaidVorRangePower

5.338 VHF Navaid VOR Range/Power

WaypointUsage

5.82 Waypoint Usage