lib_airnav::parsers::arinc424::definitions

Module identifiers

ARINC 424 Definitions - Identifiers

This module contains the identifiers for the ARINC 424 data. Identifiers are simple alphanumeric string fields. The fields may or may not be constrained to be exactly the specified length always. Conversely, some fields are minimized as needed in the spec to accomodate record layouts.

Example is 5.6 - Airport/Heliport Identifier which describes a 4-character code that identifies an airport or heliport.

Structs

LengthLimitedIdentifier

Enums

GridMora

5.143 Grid MORA

Type Aliases

AirportHeliportIdentifier

5.6 Airport/Heliport Identifier

AirwayRestrictionIdentifier

5.154 Airway Restriction Identifier

AirwayRestrictionNotes

5.163 Airway Restriction Notes

AirwayRestrictionStartEndDate

5.157 Airway Restriction Start/End Date

ApproachRouteIdentifier

5.10 Approach Route Identifier

AtaIataDesignator

5.107 ATA/IATA Designator

Callsign

5.105 Callsign

CenterFix

5.144 Center Fix

CompanyRouteIdent

5.76 Company Route Ident

ContinuationRecordNumber

5.16 Continuation Record Number

ControllingAgency

5.140 Controlling Agency

DMEIdentifier

5.38 DME Identifier

DatumCode

5.198 Datum Code

EnrouteAlternateAirportHeliport

5.148 Enroute Alternate Airport/Heliport

EnrouteRouteIdentifier

5.8(A) Enroute Route Identifier

EnrouteTransition

5.85 Enrt Transition

FirRdoIdentifier

5.190 FIR/RDO Identifier

FirUirAddress

5.151 FIR/UIR Address

FirUirIdentifier

5.116 FIR/UIR Identifier

FirUirName

5.125 FIR/UIR Name

FixIdentifier

5.13 Fix Identifier

FromToAirportHeliportFix

5.75 From/To Airport/Heliport/Fix (FROM/TO AIRPORT/HELIPORT/FIX)

GateIdentifier

5.56 Gate Identifier (GATE IDENT)

IcaoCode

5.14 ICAO Code

InboundCourseTheta

5.15 Inbound Course Theta

InitialTerminusFixOrAirport

5.194 Initial/Terminus Fix or Airport

LocalizerMlsGlsIdentifier

5.44 Localizer/MLS/GLS Identifier (LOC, MLS, GLS IDENT)

MarkerCode

5.111 Marker Code (Morse)

MarkerIdentifier

5.110 Marker Identifier (Enroute Marker) (IDENT)

MoraStartingLatitude

5.141 MORA Starting Latitude

MoraStartingLongitude

5.142 MORA Starting Longitude

MultipleCode

5.130 Multiple Code

Name

5.60 Name (NAME), Gate and Holding Pattern records

NameOfFacility

5.71 Name Field, Navaid/Airport/Heliport/Enroute Marker records

Notes

5.61 Notes (Continuation Records) (NOTES)

PadIdentifier

5.180 Pad Identifier

PathAndTermination

5.21 Path and Termination

PositionNarrative

5.189 Position Narrative

PreferredRouteIdentifier

5.8(B) Preferred Route Identifier

ProcedureName

5.139 Procedure Name

RecommendedNavaid

5.23 Recommended NAVAID

RegionCode

5.41 Region Code

RemoteFacility

5.200 Remote Facility

RestrictiveAirspaceDesignation

5.129 Restrictive Airspace Designation

RestrictiveAirspaceName

5.126 Restrictive Airspace Name

RunwayDescription

5.59 Runway Description (RUNWAY DESCRIPTION)

RunwayIdentifier

5.46 Runway Identifier (RUNWAY ID)

RunwayTransition

5.84 Runway Transition

SectorFacility

5.185 Sector Facility

SectorizationNarrative

5.186 Sectorization Narrative

SidStarApproachAirway

5.78 SID/STAR/App/AWY (S/S/A/AWY), SID/STAR/AWY (S/S/AWY)

SidStarRouteIdentifier

5.9

TerminalAlternateAirport

5.87 Terminal/Alternate Airport

TimeOfOperation

5.195 Time of Operation

ToFix

5.83 To Fix

TransitionIdentifier

5.11 Transition Identifier

VORNDBIdentifier

5.33 VOR/NDB Identifier

WaypointDescriptionCode

5.17 Waypoint Description Code

WaypointNameDescription

5.43 Waypoint Name/Description (NAME/DESC)